changed: put seasons videodatabase methods in separate file
